refactor(Summary Display): Rename folder to summary-display
#639
