+++ src/assets/wise5/directives/summary-display/summary-display.component.html
